requirements.txt
customtkinter
pyaudio
assemblyai
deepgram-sdk
python-dotenv
pyperclip
watchdog
tkinterdnd2
